pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "twag"
version = "0.1.0"
description = "Twitter aggregator for market-relevant signals"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}
authors = [
    {name = "Clifton King"},
]
keywords = ["twitter", "x", "aggregator", "market", "finance", "signals", "cli"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Financial and Insurance Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]
dependencies = [
    "click>=8.1.0",
    "anthropic>=0.40.0",
    "google-genai>=1.0.0",
    "httpx>=0.27.0",
    "python-dateutil>=2.8.0",
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "jinja2>=3.1.0",
    "python-multipart>=0.0.6",
    "tabulate>=0.9.0",
    "pydantic>=2.0",
    "rich>=13.0",
    "rich-click>=1.7",
]

[project.scripts]
twag = "twag.cli:cli"

[project.urls]
Homepage = "https://github.com/clifton/twag"
Repository = "https://github.com/clifton/twag"
Issues = "https://github.com/clifton/twag/issues"

[tool.hatch.build.targets.wheel]
packages = ["twag"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=6.0",
    "ruff>=0.9",
    "ty",
]

[tool.ruff]
target-version = "py310"
line-length = 120

[tool.ruff.lint]
select = ["E", "W", "F", "I", "UP", "B", "SIM", "RUF"]
ignore = [
    "E501",    # line too long (handled by formatter)
    "SIM105",  # contextlib.suppress — less readable for multi-line try blocks
    "SIM108",  # ternary — less readable for commented branches
    "SIM103",  # return condition directly — less readable with negation
    "SIM117",  # nested with — sometimes clearer than combined
    "RUF059",  # unused unpacked var — common in tuple-return APIs
]

[tool.ruff.lint.isort]
known-first-party = ["twag"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q --tb=short"

[tool.coverage.run]
source = ["twag"]
omit = ["twag/web/*"]

[tool.coverage.report]
show_missing = true
skip_empty = true

[tool.ty.environment]
python-version = "3.10"

[tool.ty.analysis]
allowed-unresolved-imports = ["google.*", "google.**", "anthropic.*", "anthropic.**"]

[tool.ty.rules]
possibly-missing-attribute = "ignore"

# Scripts use monkey-patching and aren't part of the library
[[tool.ty.overrides]]
include = ["scripts/**"]

[tool.ty.overrides.rules]
unresolved-attribute = "ignore"
unresolved-import = "ignore"

# Tests use dynamic patterns that confuse type checkers
[[tool.ty.overrides]]
include = ["tests/**"]

[tool.ty.overrides.rules]
unsupported-operator = "ignore"
not-subscriptable = "ignore"

# Processor — Row|dict union narrowing causes false positives
[[tool.ty.overrides]]
include = ["twag/processor/**"]

[tool.ty.overrides.rules]
invalid-argument-type = "ignore"
invalid-assignment = "ignore"
invalid-return-type = "ignore"
not-iterable = "ignore"

# Web layer — CORSMiddleware typing is incompatible with ty
[[tool.ty.overrides]]
include = ["twag/web/**"]

[tool.ty.overrides.rules]
invalid-argument-type = "ignore"
